Moved roomAvailable to ReservationLogic and added output for room type to Reservation
